Move function of conftest up one level to be shared between other tests
